# backend/docs/swagger.yaml
basePath: /
definitions:
  res.ErrorConstant:
    properties:
      code:
        type: integer
      data: {}
      error_fields:
        items:
          $ref: '#/definitions/res.Message'
        type: array
      message:
        type: string
      status:
        type: boolean
    type: object
  res.Message:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  schemas.AddToCartRequest:
    properties:
      product_variant_id:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - product_variant_id
    - quantity
    type: object
...
  schemas.CheckoutRequest:
    properties:
      address_id:
        type: string
      courier:
        description: e.g., jne, pos, tiki
        type: string
      shipping_service:
        description: e.g., REG, OKE
        type: string
    required:
    - address_id
    - courier
    - shipping_service
    type: object
...
  schemas.UpdateCartItemRequest:
    properties:
      quantity:
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: E-Commerce NW SBO Backend API
  title: E-Commerce NW API
  version: "1.0"
paths:
...
  /carts:
    delete:
      consumes:
      - application/json
      description: Remove all items from cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: DELETE Clear Cart
      tags:
      - Carts
    get:
      consumes:
      - application/json
      description: Get current logged-in user cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: GET User Cart
      tags:
      - Carts
  /carts/items:
    post:
      consumes:
      - application/json
      description: Add product variant to cart
      parameters:
      - description: Add To Cart Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.AddToCartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: POST Add To Cart
      tags:
      - Carts
  /carts/items/{id}:
    delete:
      consumes:
      - application/json
      description: Remove item from cart
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: DELETE Cart Item
      tags:
      - Carts
    put:
      consumes:
      - application/json
      description: Update quantity of cart item
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Cart Item Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: PUT Update Cart Item
      tags:
      - Carts
...
  /orders/checkout:
    post:
      consumes:
      - application/json
      description: Checkout cart items
      parameters:
      - description: Checkout Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.CheckoutRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      security:
      - BearerAuth: []
      summary: POST Checkout
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"